agent/requirements.txt
# Core dependencies
web3==6.15.1
requests==2.31.0
python-dotenv==1.0.0

# ASI Alliance / Fetch.ai
uagents==0.12.0
cosmpy==0.9.2

# Data processing
numpy==1.26.3
pandas==2.1.4

# Optional: MeTTa integration
# hyperon==0.1.11

# HTTP server for API
flask==3.0.0
flask-cors==4.0.0

